refact: extracted `notesContainer` as a component
