+++ pages/video/[id].tsx
@@ -1,27 +1,14 @@
-import {
-  Box,
-  Button,
-  FormControl,
-  Heading,
-  HStack,
-  Input,
-  Stack,
-  useColorModeValue,
-  Flex,
-  Text,
-} from "@chakra-ui/react";
+import { Box, useColorModeValue, Flex } from "@chakra-ui/react";
 import { useRouter } from "next/router";
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect, useRef } from "react";
 import VideoPlayer from "../../components/login/videoPlay";
-import { useReactToPrint } from "react-to-print";
 import axios from "axios";
 import { Navbar } from "../../components/login/navbar";
 import { useAuth } from "../../context/authContext";
-import { INote } from "../../types";
 import { ParsedUrlQuery } from "querystring";
-import { RootState } from "../../redux/store";
-import { useSelector, useDispatch } from "react-redux";
+import { useDispatch } from "react-redux";
 import { getNotes } from "../../redux/notesSlice";
+import NotesContainer from "../../components/notesContainer";
 
 type Params = {
   params: {
@@ -64,63 +51,13 @@ export async function getStaticPaths() {
 export default function Video({ id }: ParsedUrlQuery) {
   const router = useRouter();
   const { user } = useAuth();
-  const noteInput = useRef<HTMLInputElement>(null);
-  const noteState = useSelector((state: RootState) => state.notes.notesArr);
   const dispatch = useDispatch();
 
-  const componentRef = useRef<any>(null);
-
-  const ref = useRef<any>();
+  const videoRef = useRef<any>();
 
   useEffect(() => {
     dispatch(getNotes({ userId: user?._id, videoId: id }));
   }, [user]);
-
-  const handlePrint = useReactToPrint({
-    content: () => componentRef.current,
-    documentTitle: "Notevid",
-  });
-
-  function getTime(time: number) {
-    let hour, min, sec;
-    time = time;
-    hour = time / 3600;
-    time = time % 3600;
-    min = time / 60;
-    time = time % 60;
-    sec = time;
-
-    return `${Math.round(hour)}:${Math.round(min)}:${Math.round(sec)}`;
-  }
-
-  function seekTo(time: number) {
-    ref.current.seekTo(time);
-  }
-  async function addNote(e: any) {
-    if (
-      e.key === "Enter" ||
-      (e.type === "click" && noteInput.current?.value.length)
-    ) {
-      try {
-        const response = await axios.post("/api/notes", {
-          userId: user._id,
-          videoId: id,
-          note: noteInput?.current?.value,
-          timestamp: Math.round(ref.current.getCurrentTime()),
-        });
-
-        if (response.status == 201) {
-          if (noteInput.current) {
-            noteInput.current.value = "";
-
-            dispatch(getNotes({ userId: user?._id, videoId: id }));
-          }
-        }
-      } catch (err) {
-        console.log(err);
-      }
-    }
-  }
 
   return (
     <Flex
@@ -130,7 +67,6 @@ export default function Video({ id }: ParsedUrlQuery) {
       backgroundColor={useColorModeValue("gray.100", "black.100")}
     >
       <Navbar />
-      <Button onClick={handlePrint}>Print</Button>
 
       <Box
         alignItems={"center"}
@@ -139,90 +75,11 @@ export default function Video({ id }: ParsedUrlQuery) {
         height={"100vh"}
         mt={2}
       >
-        <VideoPlayer height="100vh" width="100%" id={id} reference={ref} />
+        <VideoPlayer height="100vh" width="100%" id={id} reference={videoRef} />
       </Box>
 
       <Box>
-        <Box
-          mt={2}
-          boxShadow={useColorModeValue("lg", "dark-lg")}
-          rounded={"lg"}
-          mx={"auto"}
-          minW={"30vw"}
-          py={2}
-          px={6}
-          maxH={"90vh"}
-        >
-          <Stack align={"center"}>
-            <Heading py={2} fontSize={"3xl"} textAlign={"center"}>
-              Notes 📝
-            </Heading>
-          </Stack>
-
-          <Box
-            ref={componentRef}
-            rounded={"lg"}
-            bg={useColorModeValue("light.100", "dark.100")}
-            p={8}
-            pb={1}
-            height={"75vh"}
-            maxH={"100%"}
-          >
-            <Box>
-              <Box
-                bgColor={"green"}
-                alignItems={"center"}
-                overflowY={"auto"}
-              ></Box>
-              <Box display={"flex"} flexDirection={"column"} height={"60vh"}>
-                {noteState.map((item: INote, index: number) => {
-                  return (
-                    <Box
-                      display={"flex"}
-                      justifyContent={"space-between"}
-                      key={index}
-                      my={1}
-                      width={"100%"}
-                    >
-                      <Text>{item.note}</Text>
-                      <Text
-                        _hover={{ color: "blue.100" }}
-                        fontWeight={"medium"}
-                        cursor={"pointer"}
-                        onClick={() => {
-                          seekTo(item.timestamp);
-                        }}
-                      >
-                        {getTime(item.timestamp)}
-                      </Text>
-                    </Box>
-                  );
-                })}
-              </Box>
-
-              <HStack
-                py={2}
-                mt={"auto"}
-                display={"flex"}
-                alignSelf={"flex-end"}
-              >
-                <FormControl id="email" isRequired>
-                  <Input
-                    ref={noteInput}
-                    border="1px"
-                    borderColor={useColorModeValue("dark.100", "white.100")}
-                    type="text"
-                    onKeyDown={(e) => {
-                      addNote(e);
-                    }}
-                  />
-                </FormControl>
-
-                <Button onClick={addNote}>Note</Button>
-              </HStack>
-            </Box>
-          </Box>
-        </Box>
+        <NotesContainer id={id} videoRef={videoRef} />
       </Box>
     </Flex>
   );
